+++ admin/include/tintuc/xuly.php
@@ -1,43 +1,46 @@
 <?php 
 include("../../config.php");
 session_start();
-$form = $_GET["form"];
-$id = $_GET["id"];
-$muctin = $_POST["muctin"];
-$tieude = $_POST["tieude"];
-$noidung = $_POST["noidung"];
-$tenfile = $_FILES["anh"]["name"];
-$noibat = $_POST["noibat"];
-$trangthai = $_POST["trangthai"];
+if(isset($_GET["form"])){$form = $_GET["form"];}else{$form = "";}
+if(isset($_GET["id"])){$id = $_GET["id"];}else{$id = "";}
+if(isset($_POST["muctin"])){$muctin = $_POST["muctin"];}else{$muctin = '0';}
+if(isset($_POST["tieude"])){$tieude = $_POST["tieude"];}else{$tieude = "";}
+if(isset($_POST["noidung"])){$noidung = $_POST["noidung"];}else{$noidung = "";}
+if(isset($_POST["noibat"])){$noibat = $_POST["noibat"];}else{$noibat = "0";}
+if(isset($_POST["trangthai"])){$trangthai = $_POST["trangthai"];}else{$trangthai = "";}
+if(isset($_FILES["anh"])){$tenfile = $_FILES["anh"]["name"];}else{$tenfile= "";}
 $ngaynhap = date("Y-m-d");
 $nguoinhap = $_SESSION["user_huye_name"];
 $time = date("hisdmY");
-$filename = $_GET["filename"];
+if(isset($_GET["filename"])){$filename = $_GET["filename"];}else{$filename="";}
 // xử lý tên file
 if($tenfile != ""){
     $anh = $time.$tenfile;
     $dichcopy = "../../../imguploads/".$anh;
-}else{$dichcopy="";}
+}else{
+    $anh = "";
+    $dichcopy="";
+}
 // Lấy thông tin phân quyền
     $sqlphanquyen = "select them, sua, xoa from phanquyen where user = '$_SESSION[user_huye_id]' and form = '$form'";
-    $tbphanquyen = mysql_query($sqlphanquyen);
-    $rsphanquyen = mysql_fetch_array($tbphanquyen);
+    $tbphanquyen = mysqli_query($con,$sqlphanquyen);
+    $rsphanquyen = mysqli_fetch_array($tbphanquyen);
 // thực hiện thêm tin mới
 if(isset($_POST["them"])){
     if($rsphanquyen["them"]== 1){
         if($tenfile != ""){copy($_FILES["anh"]["tmp_name"],$dichcopy);}
         $sql = "insert into tintuc(muctin, tieude, noidung, anh, noibat, trangthai, ngaynhap, nguoinhap) values('$muctin','$tieude','$noidung','$anh','$noibat','$trangthai','$ngaynhap','$nguoinhap')";
-        mysql_query($sql);
+        mysqli_query($con,$sql);
         header("location: ../../index.php?form=".$form);
     }else{
         header("location: ../../index.php?form=".$form."&false=false");
     }
-    
 }
 if(isset($_POST["sua"])){
+   
     if($rsphanquyen["sua"]== 1){
         if($tenfile != ""){
-            if(file_exists("../../../imguploads/".$filename)){
+            if(file_exists("../../../imguploads/".$filename && $filename!="")){
                 unlink("../../../imguploads/".$filename);
             }
             copy($_FILES["anh"]["tmp_name"],$dichcopy);
@@ -45,7 +48,8 @@ if(isset($_POST["sua"])){
         }else{
             $sql = "update tintuc set muctin = '$muctin', tieude = '$tieude', noidung = '$noidung',noibat = '$noibat',trangthai = '$trangthai', nguoinhap = '$nguoinhap' where id = '$id'";
         }
-        mysql_query($sql);
+        // echo $sql;
+        mysqli_query($con,$sql);
         header("location: ../../index.php?form=".$form."&act=edit&id=".$id);
     }else{
         header("location: ../../index.php?form=".$form."&false=false");
@@ -57,7 +61,7 @@ if(isset($_POST["xoa"])){
             unlink("../../../imguploads/".$filename);
         }
         $sql = "delete from tintuc where id = '$id'";
-        mysql_query($sql);
+        mysqli_query($con,$sql);
         header("location: ../../index.php?form=".$form);
     }else{
         header("location: ../../index.php?form=".$form."&false=false");
